Backend: move REST client into backend package
