+++ tests/test_api_ensure_prog.py
@@ -7,7 +7,7 @@ from unittest.mock import AsyncMock
 import aiohttp
 import pytest
 
-from custom_components.termoweb.api import RESTClient
+from custom_components.termoweb.backend.rest_client import RESTClient
 
 
 @pytest.fixture
