+++ tests/test_api_ensure_temperature.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import pytest
 
-from custom_components.termoweb.api import RESTClient
+from custom_components.termoweb.backend.rest_client import RESTClient
 from tests.test_api import FakeSession
